Don't mark local classes public
The AT generator can get confused if local classes are in a mapped class
and that local class isn't mapped, since it's package won't match the
containing class's package. This is illegal as local classes can't be
public. Local class access should never need to be transformed though,
so we can safely ignore local classe.