Remove imports when the imported name clashes with the new class name
When the name of the class to be remapped clashes with an existing
import, that import needs to be removed and all references to that class
in the file should be fully qualified instead. The import rewriter
handles that perfectly by just removing the import when we detect a
clash.