Fix overlapping MultiTextEdits causing conflicts
This takes all of the children in the `MutliTextEdit` we want to merge
and makes them children of the `MutliTextEdit` being merged into.
Conflicts were occurring when one MTE's edit range was overlapping
with the others edit range.

I ran this over multiple codebases and this change didn't affect
anything other than the static imports that were failing.